Move metrics into it's own API endpoint, and point the frontend at it
